% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustersAndLoess.R
\name{apply_LOESS}
\alias{apply_LOESS}
\title{Apply LOESS smoothing to barcode trajectories}
\usage{
apply_LOESS(c_series)
}
\arguments{
\item{c_series}{A dataframe containing \code{time}, \code{frequency}, and \code{cluster} columns.}
}
\value{
A dataframe with LOESS-smoothed values.
}
\description{
Apply LOESS smoothing to barcode trajectories
}
