% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDiversity.R
\name{format_sample}
\alias{format_sample}
\title{This function allows you to compute barcode diversity for all samples and
create dataframes with diversities per cohort}
\usage{
format_sample(sample)
}
\description{
This function allows you to compute barcode diversity for all samples and
create dataframes with diversities per cohort
}
