% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardIndex.R
\name{jaccardIndex}
\alias{jaccardIndex}
\alias{plotJaccardIndex}
\alias{pairwiseJaccard}
\title{Here we compute the Jaccard index for all pairs of sets (either top barcodes
per sample by maximum frequency or by final frequency). The Jaccard similarity
index compares members from two sets to see: which members are shared and which
are distinct. The higher the percentage, the more similar the two populations are.}
\usage{
plotJaccardIndex(topFinal, topMax, library_topN, names)

pairwiseJaccard(l1)

jaccardIndex(s1, s2)
}
\arguments{
\item{topFinal}{a list of dataframes containing barcode frequencies (sorted by final freq.) of all samples}

\item{topMax}{a list of dataframes containing barcode frequencies (sorted by max freq.) of all samples}

\item{library_topN}{a dataframe containing the top N most frequent barcodes of the initial library}

\item{names}{a list of sample names}

\item{l1}{a list containing the top N IDs of the samples and the initial library}

\item{s1}{a set of top IDs}

\item{s2}{a set of top IDs}
}
\description{
This file contains multiple functions. The main function is: plotJaccardIndex()
and it uses jaccardIndex() and pairwiseJaccard() which are all defined here.
}
