% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performHClustering.R
\name{perform_hierarchical_clustering}
\alias{perform_hierarchical_clustering}
\title{Get the clusters resulting from the hierarchical clustering of the barcoded lineages.}
\usage{
perform_hierarchical_clustering(
  filtered_data,
  agglomeration_method,
  similarity_metric,
  missing_values = NULL
)
}
\arguments{
\item{filtered_data}{a dataframe filtered by filterData()}

\item{agglomeration_method}{a character string indicating which method is to be used for agglomeration}

\item{similarity_metric}{a character string indicating which metric to be used to measure similarity between two time-series}

\item{missing_values}{a character string indicating which method is to be used to manage missing values while computing covariances}
}
\value{
A dataframe containing the resulting clusters of a hierarchical clustering
for one or multiple thresholds
}
\description{
This function uses the data filtered by filterData() to compute a distance matrix
according to pearson's or dtw's approach. From the resulting matrix, we proceed
to a hierarchical clustering of the data according to a given agglomeration method.
Then, in order to visualize the results,we plot a dendrogram and a heatmap.
Finally, if no threshold is provided, we compute the relative clusters for ALL
thresholds between 0.1 and maximum height of hierarchical clustering tree. This step allows the
user to visualize the possible clusters and make an informed choice about which
threshold to use.
}
