% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{A plot function}
\usage{
plotCorrelation(input_data_sample, initial_library, sample_name)
}
\arguments{
\item{input_data_sample}{input file given by user}

\item{initial_library}{barcode library}

\item{sample_name}{the sample's name}
}
\value{
A dataframe linking the sample to the library
}
\description{
This function allows you to create a running scatterplot of each sample's barcode
frequencies at all time points vs the barcode frequencies at time 0. In other
words, the plot shows the correlation between initial frequencies and frequencies at time t.
}
