% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInitFreq.R
\name{plotInitFreq}
\alias{plotInitFreq}
\title{A plotting function}
\usage{
plotInitFreq(initial_library)
}
\arguments{
\item{initial_library}{a dataframe containing information on the initial barcode library}
}
\value{
A dataframe containing information on the initial barcode library
}
\description{
This function allows you to plot the barcode frequencies from the initial library.
}
