% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeData.R
\name{reshapeDF}
\alias{reshapeDF}
\title{A reshape function
This function allows you to reshape data for plotting.}
\usage{
reshapeDF(raw_df)
}
\description{
A reshape function
This function allows you to reshape data for plotting.
}
