% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{theme_Matrix}
\alias{theme_Matrix}
\title{Theme for matrix-style plots (e.g. heatmaps or abundance matrices)}
\usage{
theme_Matrix(base_size = 24, base_family = "Arial")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Theme for matrix-style plots (e.g. heatmaps or abundance matrices)
}
