% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{theme_Publication_DTW}
\alias{theme_Publication_DTW}
\title{Custom ggplot2 theme for DTW alignment}
\usage{
theme_Publication_DTW(
  base_size = 32,
  base_family = "Helvetica",
  aspect.ratio = 0.75
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{aspect.ratio}{Aspect ratio of the plot.}
}
\value{
A ggplot2 theme object.
}
\description{
Custom ggplot2 theme for DTW alignment
}
