% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{theme_Publication_noYaxis}
\alias{theme_Publication_noYaxis}
\title{Custom ggplot2 theme with no y-axis title}
\usage{
theme_Publication_noYaxis(
  base_size = 24,
  base_family = "Arial",
  aspect.ratio = 0.75
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{aspect.ratio}{Aspect ratio of the plot.}
}
\value{
A ggplot2 theme object.
}
\description{
Custom ggplot2 theme with no y-axis title
}
