% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_span.R
\name{adjust_span}
\alias{adjust_span}
\title{Recursively calculates the span for a LOESS curve.}
\usage{
adjust_span(x, y, span)
}
\arguments{
\item{x}{numeric data from a dataframe for which we want to produce a LOESS curve}

\item{y}{numeric data from a dataframe for which we want to produce a LOESS curve}

\item{span}{a double representing the desired span to compute the LOESS curves}
}
\description{
In general, the distribution and quantity of data influences the value of the LOESS's span.
To avoid manipulation of the span by the user, its value is computed recursively from an initial value of 0.2.
}
