% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDiversity.R
\name{calculateDiversity}
\alias{calculateDiversity}
\alias{calculate_diversity}
\alias{format_sample}
\alias{calculate_q_0}
\alias{calculate_q_1}
\alias{calculate_q_inf}
\title{Compute barcode diversity for all samples and create dataframes with
diversities per cohort.}
\usage{
calculate_diversity(data)

format_sample(sample)

calculate_q_0(mat)

calculate_q_1(mat)

calculate_q_inf(mat)
}
\arguments{
\item{data}{reshaped data produced by format_sample()}

\item{sample}{input dataframe given by user}

\item{mat}{matrix containing number of barcodes for each ID over each timepoint}
}
\value{
A dataframe containing all diversity indices for a sample.
}
\description{
This file contains multiple functions. The main function is: calculate_diversity()
and it uses format_sample(), calculate_q_0(), calculate_q_1() and calculate_q_inf()
which are all defined here. While calculate_q_0(), calculate_q_1() and calculate_q_inf()
computes the diversity index for different q-values, calculate_diversity() binds the indices
together.
}
