% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDiversity.R
\name{calculateDiversity}
\alias{calculateDiversity}
\alias{calculate_diversity}
\alias{format_sample}
\alias{calculate_q_0}
\alias{calculate_q_1}
\alias{calculate_q_inf}
\title{A function to calculate a sample's diversity}
\usage{
calculate_diversity(input_data)

format_sample(sample)

calculate_q_0(mat)

calculate_q_1(mat)

calculate_q_inf(mat)
}
\arguments{
\item{input_data}{input dataframe given by user}

\item{sample}{input dataframe given by user}

\item{mat}{matrix containing the number of barcodes (READS) for each ID over each time-point}
}
\value{
A dataframe containing all diversity indices for a sample.
}
\description{
Here, we compute the input's barcode diversity.
This file contains multiple functions. The main function is: calculate_diversity()
and it uses format_sample(), calculate_q_0(), calculate_q_1() and calculate_q_inf().
While calculate_q_0(), calculate_q_1() and calculate_q_inf() compute the diversity
index for different q-values, calculate_diversity() binds the indices together.
}
