% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSpearmanCorr.R
\name{calculateSpearmanCorr}
\alias{calculateSpearmanCorr}
\title{Here we calculate the running frequency correlations with initial library
for all samples.}
\usage{
calculateSpearmanCorr(index_dataframe, sample_name)
}
\arguments{
\item{index_dataframe}{dataframe returned by plotCorrelation()}

\item{sample_name}{the sample's name}
}
\value{
A dataframe containing the running frequency correlations between a
sample and the initial library for all timepoints
}
\description{
Here we calculate the running frequency correlations with initial library
for all samples.
}
