% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTop.R
\name{fetchTop}
\alias{fetchTop}
\title{A fetch function
This function allows you to fetch the top N barcodes per sample.}
\usage{
fetchTop(reshaped_df, n_intersect)
}
\description{
A fetch function
This function allows you to fetch the top N barcodes per sample.
}
