% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTop.R
\name{fetchTop}
\alias{fetchTop}
\title{A fetching function}
\usage{
fetchTop(reshaped_df, N_LINEAGES)
}
\arguments{
\item{reshaped_df}{a dataframe returned by reshapeData()}

\item{N_LINEAGES}{integer representing the number of barcodes to fetch}
}
\value{
top N barcodes according to their maximum frequencies.
}
\description{
This function allows you to retrieve the top N_LINEAGES barcodes.
The top N_LINEAGES barcodes are recovered according to their maximum frequencies.
}
