% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filtering the input files}
\usage{
filterData(input_df, freq_treshold, time_threshold, name)
}
\arguments{
\item{input_df}{input file processed by read_csv()}

\item{freq_treshold}{a double representing a minimum mean frequency below which
the data is not retained.}

\item{time_threshold}{an integer indicating the time-point threshold (keep barcodes with at least x non-zero time points)}

\item{name}{a character string indicating the sample's name}
}
\value{
a matrix containing the ID,
frequency at each time point, average frequency and duration of each remaining barcode
}
\description{
This function is used to clean the input data given by the user in order to perform
the clustering. We set a minimum mean frequency and a minimum time-point threshold below which
the data is not retained. The function writes a csv file containing the ID,
frequency at each time point, average frequency and duration of each remaining barcode.
}
