% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filtering the input file}
\usage{
filterData(input_df, freq_threshold, time_threshold)
}
\arguments{
\item{input_df}{input file given by the user}

\item{time_threshold}{an integer indicating the time-point threshold (i.e. we keep the lineages with at least "time_threshold" non-zero time points)}

\item{freq_treshold}{a double representing a minimum mean frequency below which
the data is not retained for clustering.}
}
\value{
a matrix containing the ID,
frequency at each time point, average frequency and duration of each remaining barcode
}
\description{
This function is used to clean the input data given by the user in order to perform
the clustering. We set a minimum mean frequency and a minimum time-point threshold below which
the data is not retained for clustering. The function writes a csv file containing the ID,
frequency at each time point, average frequency and duration of each remaining barcode.
}
