% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterHC.R
\name{filterHC}
\alias{filterHC}
\title{Filtering the hierarchical clusters}
\usage{
filterHC(series_filtered, clusters, n_members)
}
\arguments{
\item{series_filtered}{a dataframe filtered by filterData()}

\item{clusters}{a dataframe containing the clusters from a hierarchical clustering
for one or multiple thresholds}

\item{n_members}{an integer indicating the minimum number of members per cluster}
}
\value{
A list containing a dataframe with the resulting filtered clusters, and
the minimum average frequency that must be reached by at least
one of the lineages of potentially ignored clusters for them to be taken into account
}
\description{
Once the hierarchical clustering is done, we keep only the clusters with at least "n_members" members.
By ignoring clusters with fewer than"n_members" members, we are potentially ignoring dominant clusters.
To prevent this, we ask the user to provide a minimum average frequency that must be reached by at least
one of the lineages of potentially ignored clusters for them to be taken into account.
}
