% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performHClustering.R
\name{perform_hierarchical_clustering}
\alias{perform_hierarchical_clustering}
\title{Get the clusters resulting from a hierarchical clustering of the clonal lineages.}
\usage{
perform_hierarchical_clustering(
  filtered_data,
  sample_name,
  linkage_method = c("average", "ward.D", "ward.D2", "centroid", "single", "complete",
    "median", "mcquitty"),
  clustering_method = c("pearson", "dtw"),
  covariance = NULL,
  threshold = NULL
)
}
\arguments{
\item{filtered_data}{a dataframe filtered by filterData()}

\item{sample_name}{a character string indicating the sample's name}

\item{linkage_method}{a character string indicating which method is to be used for linkage}

\item{clustering_method}{a character string indicating which method is to be used for clustering}

\item{covariance}{a character string indicating which method is to be used to compute the covariance}

\item{threshold}{an optional parameter to specify which threshold to use}
}
\value{
A dataframe containing the clusters from a hierarchical clustering
for one or multiple thresholds
}
\description{
This function uses the data filtered by filterData() to compute a distance matrix
according to pearson's or dtw's approach. From the resulting matrix, we proceed
to a hierarchical clustering of the data according to a given linkage method.
Then, in order to visualize the results,we plot a dendrogram and a heatmap.
Finally, if no threshold is provided, we compute the relative clusters for ALL
thresholds between 0.1 and max height of hierarchical clustering. This step allows the
user to visualize the possible clusters and make an informed choice about which
threshold to use.If a threshold is provided, the function returns a dataframe
containing the clusters related to that threshold.
}
