% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustersAndLoess.R
\docType{data}
\name{plotClustersAndLoess}
\alias{plotClustersAndLoess}
\alias{cluster.colors}
\alias{plotClusterLog10}
\alias{apply_LOESS}
\title{Plot the log10-transformed barcode frequencies and the moving averages (LOESS)}
\format{
An object of class \code{character} of length 26.
}
\usage{
cluster.colors

plotClusterLog10(df, cluster, sample, color, tf, effective.breaks)

apply_LOESS(series_reshaped, effective.breaks, sample_name)
}
\arguments{
\item{effective.breaks}{a list containing the breaks for the plots}

\item{sample_name}{a character string indicating the sample's name}

\item{series.filtered}{a dataframe filtered by filterData()}

\item{selected.clusters}{A dataframe containing the clusters from a hierarchical clustering
for a specific threshold}

\item{n_members}{an integer indicating the minimum number of members per cluster}
}
\description{
This file contains multiple functions. The main function is: plot_clusters_and_loess()
and it uses plotClusterLog10() and applyLOESS(). plot_clusters_and_loess() starts in
the same way as plotHCQuantification(), but without a cutoff column. In plot_clusters_and_loess(),
we plot the log10-transformed barcode frequencies contained in all selected clusters,
we compute a moving average per cluster and group them in a plot. We also write the files associated with
those two plots.
}
\keyword{datasets}
