% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustersAndLoess.R
\docType{data}
\name{plotClustersAndLoess}
\alias{plotClustersAndLoess}
\alias{cluster.colors}
\alias{plotClusterLog10}
\alias{apply_LOESS}
\title{Plot the log10-transformed barcode frequencies and the moving averages (LOESS)}
\format{
An object of class \code{character} of length 26.
}
\usage{
cluster.colors

plotClusterLog10(df, cluster, color, tf, effective.breaks)

apply_LOESS(c_series)
}
\arguments{
\item{selected_clusters}{A dataframe containing the clusters from a hierarchical clustering
for a specific threshold}
}
\description{
This file contains multiple functions. The main function is: plot_clusters_and_loess()
and it uses plotClusterLog10() and apply_LOESS(). In plot_clusters_and_loess(),
we plot the log10-transformed barcode frequencies contained in all selected clusters,
we compute a moving average per cluster and group them in a plot. We also write the files associated with
these two plots.
}
\keyword{datasets}
