% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustersAndLoess.R
\name{plotClustersAndLoess}
\alias{plotClustersAndLoess}
\alias{plot_clusters_and_loess}
\alias{plotClusterLog10}
\title{Plot the log10-transformed barcode frequencies and the moving averages (LOESS)}
\usage{
plot_clusters_and_loess(
  series.filtered,
  selected.clusters,
  sample_name,
  effective.breaks
)

plotClusterLog10(df, cluster, sample, color, tf, effective.breaks)
}
\arguments{
\item{series.filtered}{a dataframe filtered by filterData()}

\item{selected.clusters}{A dataframe containing the clusters from a hierarchical clustering
for a specific threshold}

\item{sample_name}{a character string indicating the sample's name}

\item{effective.breaks}{a list containing the breaks for the plots}
}
\description{
This file contains multiple functions. The main function is: plot_clusters_and_loess()
and it uses plotClusterLog10(). plot_clusters_and_loess() starts in
the same way as plotHCQuantification(), but without a cutoff column. In plot_clusters_and_loess(),
we plot the log10-transformed barcode frequencies contained in all selected clusters,
we compute a moving average per cluster and group them in a plot. We also write the files associated with
those two plots.
}
\details{
(write details for plotClusterLog10())
}
