% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiversity.R
\name{plotDiversity}
\alias{plotDiversity}
\title{A plot function
This function allows you to plot barcode diversities for each sample.
Each panel of each cohort is saved individually. Then we save a multi-panel
figure with all diversities and the legend.}
\usage{
plotDiversity(dataframe, cohort_names)
}
\description{
A plot function
This function allows you to plot barcode diversities for each sample.
Each panel of each cohort is saved individually. Then we save a multi-panel
figure with all diversities and the legend.
}
