% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiversity.R
\name{plotDiversity}
\alias{plotDiversity}
\title{A function to plot the diversity dynamics}
\usage{
plotDiversity(dataframe, cohort_names)
}
\arguments{
\item{dataframe}{a dataframe containing the barcode diversities for each sample}

\item{cohort_names}{the names of each sample}
}
\value{
A plot of the diversity over time.
}
\description{
This function allows you to plot barcode diversities for each sample. We save
a multi-panel figure with all diversities and the legend.
}
