% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDynamics.R
\name{plotDynamics}
\alias{plotDynamics}
\title{A plot function
This function allows you to plot barcode dynamics (Muller-style area plot and
log-line plot)}
\usage{
plotDynamics(raw_df, freq_treshold, cohort_name)
}
\description{
A plot function
This function allows you to plot barcode dynamics (Muller-style area plot and
log-line plot)
}
