% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDynamics.R
\name{plotDynamics}
\alias{plotDynamics}
\title{A plot function}
\usage{
plotDynamics(reshaped_df, colored_topFreq_df, min_freq_threshold, plot_model)
}
\arguments{
\item{reshaped_df}{a dataframe produced by reshapeData()}

\item{colored_topFreq_df}{a dataframe containing the top barcodes with their respective colors}

\item{min_freq_threshold}{a double representing the minimum frequency above which barcodes are assigned colors}

\item{plot_model}{a character string indicating what kind of plot the user wants to model}
}
\description{
This function allows you to plot barcode dynamics (logarithmic-scale and/or linear-scale).
}
