% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDynamics.R
\name{plotDynamics}
\alias{plotDynamics}
\title{A plot function}
\usage{
plotDynamics(raw_df, freq_threshold, cohort_name, plot_model)
}
\arguments{
\item{raw_df}{a dataframe produced by reshapeDF()}

\item{freq_threshold}{a double representing a limit frequency above which barcodes have assigned colors}

\item{cohort_name}{a character string indicating the sample's name}

\item{plot_model}{a character string indicating what kind of plot the user wants to model}
}
\description{
This function allows you to plot barcode dynamics (linear plot and
log-transformed plot).
}
