% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHCQuantification.R
\name{plotHCQuantification}
\alias{plotHCQuantification}
\alias{melt_dist}
\alias{applyLOESS}
\title{Plot hierarchical clustering quantification}
\usage{
plotHCQuantification(clusters_filtered)

melt_dist(dist, order = NULL, dist_name = "dist")

applyLOESS(clusters_filtered)
}
\arguments{
\item{clusters_filtered}{a dataframe filtered by filterHC()}
}
\description{
This file contains multiple functions. The main function is plotHCQuantification()
and it uses melt_dist(), applyLOESS() and adjust_span(). plotHCQuantification() is called right after perform_hierarchical_clustering()
and uses the resulting dataframe. First, we compute the moving average of barcode
frequencies for each cluster USING LOESS. The moving average is computed using only
the persistent barcodes and each cluster must contain at least N members. Then,
we calculate the smallest distance between clones (for all thresholds). We plot
a graph representing the smallest distance between two clusters and the number of clusters
according to the threshold.
}
