% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHCQuantification.R
\name{plotHCQuantification}
\alias{plotHCQuantification}
\alias{melt_dist}
\title{Plot hierarchical clustering quantification}
\usage{
plotHCQuantification(filtered_data, sample_name, clusters)

melt_dist(dist, order = NULL, dist_name = "dist")
}
\arguments{
\item{filtered_data}{a dataframe filtered by filterData()}

\item{sample_name}{a character string indicating the sample's name}

\item{clusters}{a dataframe containing the clusters from a hierarchical clustering
for one or multiple thresholds}
}
\description{
This file contains multiple functions. The main function is plotHCQuantification()
and it uses melt_dist(). plotHCQuantification() is called right after get_clonal_lineage_clusters()
and uses the resulting dataframe. First, we compute the moving average of barcode
frequencies for each cluster USING LOESS. The moving average is computed using only
the persistent barcodes and each cluster must contain at least 8 members. Then,
we calculate the smallest distance between clones (for all thresholds). We plot
a graph representing the smallest distance between two clusters and the number of clusters
according to the threshold.
}
