% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpearmanCorr.R
\name{plotSpearmanCorr}
\alias{plotSpearmanCorr}
\title{Here we plot the running frequency correlations with initial library
for all samples.}
\usage{
plotSpearmanCorr(sample_corr)
}
\arguments{
\item{sample_corr}{a dataframe containing each sample's Spearman correlation over time}
}
\value{
A plot of the Spearman rank correlation of all samples over time
}
\description{
Here we plot the running frequency correlations with initial library
for all samples.
}
