% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeData.R
\name{reshapeData}
\alias{reshapeData}
\title{A reshape function}
\usage{
reshapeData(input_data)
}
\arguments{
\item{input_data}{a dataframe with ID, Time and Reads columns.}
}
\value{
A dataframe containing the max, start, final and average frequencies of each ID in decreasing order of maximum frequency.
}
\description{
This function allows you to reshape your data for further manipulation.
The given input is a format with ID, Time and Reads columns.
We reshape it to obtain a dataframe containing the max, start, final and average frequencies of each ID.
}
